One more adjustment to TestModel

+++ Core/Tests/MVVM/TestModel.swift
@@ -2,7 +2,7 @@ import Pilot
 
 struct TM: Model {
     // swiftlint:disable:next variable_name
-    init(id: ModelId, version: Int64) {
+    init(id: ModelId, version: Int) {
         self.modelId = id
         self.modelVersion = ModelVersion(hash: version)
     }
